Account folder moved to Features folder
