+++ src/Application/Features/Account/Commands/LogoutAllCommandHandler.cs
@@ -1,22 +1,19 @@
 using Application.Common;
 using Application.Common.Exceptions;
-using Application.Common.Interfaces.Contexts;
 using Application.Common.Interfaces.Services;
 using Application.Contract.Account.Commands;
 using Domain.Entities;
 using MediatR;
 using Microsoft.AspNetCore.Identity;
-using Microsoft.EntityFrameworkCore;
 
-namespace Application.Account.Commands;
+namespace Application.Features.Account.Commands;
 
-public sealed class ChangeMyPasswordCommandHandler(
-    IApplicationDbContext db,
+public sealed class LogoutAllCommandHandler(
     UserManager<ApplicationUser> userManager,
     ICurrentUserService currentUser)
-    : IRequestHandler<ChangePasswordCommand, Unit>
+    : IRequestHandler<LogoutAllCommand, Unit>
 {
-    public async Task<Unit> Handle(ChangePasswordCommand request, CancellationToken ct)
+    public async Task<Unit> Handle(LogoutAllCommand request, CancellationToken ct)
     {
         var userName = currentUser.GetType().GetMethod("GetUserName")?.Invoke(currentUser, null) as string
                        ?? (currentUser.GetType().GetProperty("UserName")?.GetValue(currentUser) as string);
@@ -24,14 +21,12 @@ public sealed class ChangeMyPasswordCommandHandler(
         if (string.IsNullOrWhiteSpace(userName))
             throw new UnauthorizedAccessException("Пользователь не распознан.");
 
-        // Держим все операции в рамках UserManager (без db.SaveChanges)
         var user = await userManager.FindByNameAsync(userName)
                    ?? throw new NotFoundException("Пользователь не найден.");
 
-        var res = await userManager.ChangePasswordAsync(user, request.CurrentPassword, request.NewPassword);
+        var res = await userManager.UpdateSecurityStampAsync(user);
         res.ThrowBadRequestIfError();
 
-        await userManager.UpdateSecurityStampAsync(user);
         return Unit.Value;
     }
 }
