+++ src/Application/Features/Account/Commands/UpdateMyProfileCommandHandler.cs
@@ -7,7 +7,7 @@ using MediatR;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 
-namespace Application.Account.Commands;
+namespace Application.Features.Account.Commands;
 
 public sealed class UpdateMyProfileCommandHandler(
     IApplicationDbContext db,
