+++ src/Application/Features/Account/Queries/GetMyJwtQueryHandler.cs
@@ -9,7 +9,7 @@ using MediatR;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 
-namespace Application.Account.Queries;
+namespace Application.Features.Account.Queries;
 
 public sealed class GetMyJwtQueryHandler(
     IApplicationDbContext db,
